+++ app/dashboard/layout.tsx
@@ -1,4 +1,3 @@
-import MealsTable from "@/components/MealsTable/MealsTable";
 import { Card } from "@/components/ui/card";
 import React from "react";
 
@@ -13,11 +12,7 @@ const DashboardLayout = ({
 }>) => {
   return (
     <div className="flex flex-col lg:flex-row gap-4 justify-around w-full p-2 md:p-4">
-      {/* <div className="flex lg:flex-1 gap-6 flex-col lg:flex-row">
-        {calendar}
-        {children}
-      </div>
-      <div className="lg:flex-1">{table}</div> */}
+     
       <div className="w-full lg:w-3/4 flex flex-col gap-4 order-2 lg:order-1">
         <div className="flex flex-col lg:flex-row gap-4 ">
           {children}
