+++ app/ingredients/page.tsx
@@ -1,6 +1,3 @@
-import ContainerLayout from "@/components/ContainerLayout/ContainerLayout";
-import CreateMealsOrIngredients from "@/components/CreateMealsOrIngredients/CreateMealsOrIngredients";
-import CreatedMealsAndIngredients from "@/components/Dashboard/CreatedMealsAndIngredients/CreatedMealsAndIngredients";
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
 import React from "react";
@@ -13,8 +10,7 @@ const page = async () => {
   return (
    
       <div className="bg-white w-full h-screen rounded-3xl grid lg:grid-cols-2 lg:p-12 gap-6">
-      <CreateMealsOrIngredients />
-      <CreatedMealsAndIngredients type='full'/>
+    
       </div>
 
   );
