+++ components/Dashboard/CalendarView.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useState } from "react";
+import React from "react";
 import { Calendar } from "@/components/ui/calendar";
 import { useCalendarStore } from "@/store/store";
