+++ components/Dashboard/MealsTable/MealsTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
+import { Table, TableHead, TableHeader, TableRow } from "@/components/ui/table";
 import { getServerSession } from 'next-auth';
 import MealsTableBody from './MealsTableBody/MealsTableBody';
