+++ components/Dashboard/MealsTable/MealsTableBody/MealsTableRow/MealsTableRow.tsx
@@ -2,9 +2,9 @@
 import { TableCell, TableRow } from "@/components/ui/table";
 import { GetMealsResponse } from "@/lib/getMeals/getMeals";
 import { formattedDate } from "@/lib/utils";
-import { useCalendarStore, useTotalNutritionStore } from "@/store/store";
+import { useCalendarStore} from "@/store/store";
 import { Session } from "next-auth";
-import React, { useEffect } from "react";
+import React from "react";
 import DeleteRow from "./DeleteRow";
 
 type MealsTableRowProps = {
