--- components/Navbar/Profile/Dropdown/Dropdown.tsx
@@ -1,22 +1,0 @@
-import React from 'react'
-import { DropdownMenuCheckboxItemProps } from "@radix-ui/react-dropdown-menu"
- 
-import { Button } from "@/components/ui/button"
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
-
-
-
-const Dropdown = () => {
-  return (
-    <div>Dropdown</div>
-  )
-}
-
-export default Dropdown
